Move activity to new package -> patientFacilityTransfer
